moved CPU Opponent related classes into the new class library
