+++ CPU-Opponent/CPUOpponent.cs
@@ -3,8 +3,9 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using TTT_Library;
 
-namespace TTT_Library
+namespace CPU_Opponent
 {
     public class CPUOpponent
     {
